requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==1.4.53
pydantic-settings==2.1.0
pymysql==1.1.0
mysql-connector-python
numpy==1.26.4
torch==2.0.1
transformers==4.30.2
tokenizers==0.13.3
sentencepiece==0.1.99
pandas
konlpy
scikit-learn
ipykernel
python-dotenv==1.0.0
httpx==0.27.2
cryptography>=3.0.0
pytest==7.4.4
Faker